requirements.txt
pandas~=1.3.4
python-dotenv~=0.19.1